requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.6.2
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.45.0
django-core==1.4.1
django-environ==0.4.5
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
environs==9.3.2
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.9.2
lazy-object-proxy==1.6.0
marshmallow==3.13.0
mccabe==0.6.1
numpy==1.21.0
pandas==1.3.0
postgres==3.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.9.3
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.7
whitenoise==5.2.0
wrapt==1.12.1
